requirements.txt
gradio[mcp]>=4.44.0
python-dotenv>=1.0.0